\name{ANES1968}
\alias{ANES1968}
\docType{data}
\title{1968 American National Election Study}
\description{
1968 American National Election Study feeling thermometers for candidates and other politically relevant figures. 
}
\usage{data("ANES1968")}
\format{
  A data frame with 1673 observations on the following 14 variables.
  \describe{
    \item{\code{Wallace}}{a numeric vector}
    \item{\code{Humphrey}}{a numeric vector}
    \item{\code{Nixon}}{a numeric vector}
    \item{\code{McCarthy}}{a numeric vector}
    \item{\code{Reagan}}{a numeric vector}
    \item{\code{Rockefeller}}{a numeric vector}
    \item{\code{Johnson}}{a numeric vector}
    \item{\code{Romney}}{a numeric vector}
    \item{\code{Kennedy}}{a numeric vector}
    \item{\code{Muskie}}{a numeric vector}
    \item{\code{Agnew}}{a numeric vector}
    \item{\code{LeMay}}{a numeric vector}
    \item{\code{vote.turnout}}{a numeric vector}
    \item{\code{presidential.vote}}{a numeric vector}
  }
}
\examples{
data(ANES1968)
}
\keyword{datasets}
