\name{bayesunfold-package}
\alias{bayesunfold-package}
\alias{bayesunfold}
\docType{package}
\title{
\packageTitle{bayesunfold}
}
\description{
\packageDescription{bayesunfold}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bayesunfold}
\packageIndices{bayesunfold}
Performs Bayesian multidimensional unfolding analysis using L-BFGS derived configuration for starting values. }
\author{
\packageAuthor{bayesunfold}

Maintainer: \packageMaintainer{bayesunfold}
}
\keyword{ package }
