\name{bu.Preprocess}
\alias{bu.Preprocess}
\title{Preprocessing for Bayesian Unfolding}
\description{
Pre-processes data to get it ready for the Bayesian Unfolding algorithm.}
\usage{
bu.Preprocess(input, ...)
}
\arguments{
  \item{input}{A matrix or data frame of values to be unfolded.}
  \item{\dots}{Currently not implemented}
}
\value{
A list of values that will be handed to the L-BFGS routine. 
}
